authとの結合処理を行った．
