+++ api/api/models/database_models.py
@@ -70,9 +70,9 @@ class Photo(Base):
     event = relationship("Event", back_populates="photo")
 
 # authとの結合時に使うところ
-# class User(Base):
-#     __tablename__ = "users"
-#     id = Column(Integer, primary_key=True, index=True)
-#     username = Column(String, unique=True, index=True)
-#     email = Column(String, unique=True, index=True)
-#     hashed_password = Column(String)
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, index=True)
+    username = Column(String, unique=True, index=True)
+    email = Column(String, unique=True, index=True)
+    hashed_password = Column(String)
