+++ api/api/routers/auth.py
@@ -1,38 +1,55 @@
+# # main.py
+# from fastapi import Depends, FastAPI, HTTPException, status
+# from fastapi import APIRouter
+# from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+# from sqlalchemy.orm import Session
+# from jose import JWTError, jwt
+# from datetime import timedelta
+
+# from ..models.auth import SessionLocal, engine,Base
+# from ..schemes.auth import UserCreate, UserUpdate, User
+# from ..cruds.auth import get_user_by_username, get_user_by_email, create_user,update_user,delete_user
+# from ..lib.auth.auth_utils import verify_password
+# from ..lib.auth.token_utils import create_access_token, decode_access_token,add_token_to_blacklist,is_token_in_blacklist,ACCESS_TOKEN_EXPIRE_MINUTES
+# # データベースの初期化
+# from dotenv import load_dotenv
+# import os
+# # .envファイルを読み込む
+# load_dotenv()
+# # 環境変数の取得
+# database_url = os.getenv('DATABASE_URL')
+# print(database_url)
+# if not os.path.exists(database_url):
+#     Base.metadata.create_all(bind=engine)
+
+# #app = FastAPI()
+# router = APIRouter()
+
+# oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")
+
+# def get_db():
+#     db = SessionLocal()
+#     try:
+#         yield db
+#     finally:
+#         db.close()
+
 # main.py
-from fastapi import Depends, FastAPI, HTTPException, status
-from fastapi import APIRouter
+from fastapi import Depends, HTTPException, status, APIRouter
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session
-from jose import JWTError, jwt
+from jose import JWTError
 from datetime import timedelta
+from api.schemes.auth import UserCreate, UserUpdate, User
+from api.cruds.auth import get_user_by_username, get_user_by_email, create_user,update_user,delete_user
+from api.lib.auth.auth_utils import verify_password
+from api.lib.auth.token_utils import create_access_token, decode_access_token,add_token_to_blacklist,is_token_in_blacklist,ACCESS_TOKEN_EXPIRE_MINUTES
+from api.database import get_db
 
-from ..models.auth import SessionLocal, engine,Base
-from ..schemes.auth import UserCreate, UserUpdate, User
-from ..cruds.auth import get_user_by_username, get_user_by_email, create_user,update_user,delete_user
-from ..lib.auth.auth_utils import verify_password
-from ..lib.auth.token_utils import create_access_token, decode_access_token,add_token_to_blacklist,is_token_in_blacklist,ACCESS_TOKEN_EXPIRE_MINUTES
-# データベースの初期化
-from dotenv import load_dotenv
-import os
-# .envファイルを読み込む
-load_dotenv()
-# 環境変数の取得
-database_url = os.getenv('DATABASE_URL')
-print(database_url)
-if not os.path.exists(database_url):
-    Base.metadata.create_all(bind=engine)
 
-#app = FastAPI()
 router = APIRouter()
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
 
 @router.post("/auth/users/", response_model=User)
 def register_user(user: UserCreate, db: Session = Depends(get_db)):
@@ -102,7 +119,7 @@ async def read_users_me(current_user: User = Depends(get_current_user)):
 
 # 更新エンドポイントの追加
 @router.put("/auth/users/update/{user_id}", response_model=User)
-async def update_user_endpoint(user_id: int, user: UserUpdate, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
+async def update_user_endpoint(user_id: int, user: UserUpdate, db: Session = Depends(get_db)):
     db_user = update_user(db, user_id, user)
     if not db_user:
         raise HTTPException(status_code=404, detail="User not found")
@@ -110,7 +127,7 @@ async def update_user_endpoint(user_id: int, user: UserUpdate, db: Session = Dep
 
 # 削除エンドポイントの追加
 @router.delete("/auth/users/delete/{user_id}", response_model=User)
-async def delete_user_endpoint(user_id: int, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
+async def delete_user_endpoint(user_id: int, db: Session = Depends(get_db)):
     db_user = delete_user(db, user_id)
     if not db_user:
         raise HTTPException(status_code=404, detail="User not found")
