+++ api/api/schemes/auth.py
@@ -18,4 +18,4 @@ class User(BaseModel):
     email: str
 
     class Config:
-        orm_mode = True
+        from_attributes = True
